[PE-305] display after registration only
